+++ src/data_row__json/index.test.ts
@@ -1,7 +1,7 @@
 import { test } from 'uvu'
 import { equal } from 'uvu/assert'
-import { data_row_, data_row__json_, type dehydrated_json_val_T, header_row_ } from '../index.js'
-test('data_row__json_|-dehydrated_json_val_', ()=>{
+import { data_row__new, data_row__json__new, type dehydrated_json_val_T, header_row_ } from '../index.js'
+test('data_row__json__new|-dehydrated_json_val_', ()=>{
 	const header_row = header_row_([
 		'col0',
 		'col1',
@@ -9,14 +9,14 @@ test('data_row__json_|-dehydrated_json_val_', ()=>{
 		'col3',
 		'col4'
 	])
-	const data_row = data_row_(
+	const data_row = data_row__new(
 		['foo', 'bar', 'baz', 1.23, new Date('2014-5-5')],
 		header_row)
 	equal(
-		data_row__json_(data_row),
+		data_row__json__new(data_row),
 		JSON.stringify(['foo', 'bar', 'baz', 1.23, new Date('2014-5-5')]))
 })
-test('data_row__json_|+dehydrated_json_val_', ()=>{
+test('data_row__json__new|+dehydrated_json_val_', ()=>{
 	const header_row =
 		header_row_<[
 			['col0', string],
@@ -32,11 +32,11 @@ test('data_row__json_|+dehydrated_json_val_', ()=>{
 			'col4'
 		])
 	const data_row =
-		data_row_(
+		data_row__new(
 			['foo', 'bar', 'baz', 1.23, new Date('2014-5-5')],
 			header_row)
 	equal(
-		data_row__json_<[
+		data_row__json__new<[
 			['col0', string],
 			['col1', string],
 			['col2', string],