+++ src/header_row/index.js
@@ -2,15 +2,16 @@ import { idx_a_ } from '@ctx-core/array'
 import { isArray, nullish__none_ } from '@ctx-core/function'
 import { computed_ } from '@ctx-core/nanostores'
 import { be_, clone, is_ctx_ } from '@ctx-core/object'
-import { table_, table__, table__new, table__set } from '../table/index.js'
+import { table_, table$_, table__new, table__set } from '../table/index.js'
 /** @typedef {import('@ctx-core/object').Ctx}Ctx */
 /** @typedef {import('../_types').column_a_T}column_a_T */
 /** @typedef {import('../_types').header_row_T}header_row_T */
-/** @type {typeof import('./index.d.ts').header_row__} */
-export const header_row__ = be_('header_row__', ctx=>
-	computed_(table__(ctx), table=>
+/** @type {typeof import('./index.d.ts').header_row$_} */
+export const header_row$_ = be_('header_row$_', ctx=>
+	computed_(table$_(ctx), table=>
 		nullish__none_([table], ()=>
 			table.header_row)))
+export { header_row$_ as header_row__ }
 /**
  * @param {Ctx|column_a_T|number}ctx_or_column_a_or_length
  * @returns {column_a_T}
@@ -20,7 +21,7 @@ export function header_row_(
 	ctx_or_column_a_or_length
 ) {
 	if (is_ctx_(ctx_or_column_a_or_length)) {
-		return header_row__(/** @type {Ctx} */ctx_or_column_a_or_length).$
+		return header_row$_(/** @type {Ctx} */ctx_or_column_a_or_length).$
 	}
 	return header_row__new(ctx_or_column_a_or_length)
 }